+++ R/aggregate.R
@@ -50,7 +50,7 @@ polygon_to_raster <- function(polygon, field = NULL) {
     }
 
     # ensure that field is numerical
-    polygon$poly_i <- seq_len(nrow(unique(polygon[[field]])))
+    polygon$poly_i <- seq_len(nrow(polygon))
     poly_rast <- terra::rasterize(x = polygon, r, field = "poly_i")
 
     poly_ID_vector <- polygon[[field]][, 1]
@@ -372,11 +372,40 @@ setMethod(
         calculateOverlap(
             x = x,
             y = y@raster_object,
-            name_overlap = objName(y),
+            name_overlap = name_overlap %null% objName(y),
             poly_subset_ids = poly_subset_ids,
             verbose = verbose,
             ...
         )
+    }
+)
+
+# * giottoPolygon giottoAffineImage ####
+#' @rdname calculateOverlap
+#' @export
+setMethod(
+    "calculateOverlap", signature(x = "giottoPolygon", y = "giottoAffineImage"),
+    function(x, y,
+        name_overlap = NULL,
+        poly_subset_ids = NULL,
+        return_gpolygon = TRUE,
+        verbose = TRUE,
+        ...
+    ) {
+        aff <- y@affine
+        res <- calculateOverlap(
+            x = affine(x, aff, inv = TRUE),
+            y = y@raster_object,
+            name_overlap = name_overlap %null% objName(y),
+            poly_subset_ids = poly_subset_ids,
+            verbose = verbose,
+            ...
+        )
+        x@overlaps <- res@overlaps
+        if (is.null(centroids(x))) {
+            x <- centroids(x, append_gpolygon = TRUE)
+        }
+        x
     }
 )
 
@@ -394,7 +423,7 @@ setMethod(
         verbose = TRUE,
         ...) {
         if (is.null(name_overlap)) {
-            .gstop("calculateOverlap: name_overlap must be given")
+            stop("calculateOverlap: name_overlap must be given", call. = FALSE)
         }
 
         res <- calculateOverlap(
@@ -410,7 +439,6 @@ setMethod(
             if (is.null(centroids(x))) {
                 x <- centroids(x, append_gpolygon = TRUE)
             }
-
             x@overlaps[["intensity"]][[name_overlap]] <- res
             return(x)
         } else {
@@ -440,7 +468,13 @@ setMethod(
         checkmate::assert_true(terra::is.polygons(x))
         GiottoUtils::package_check("exactextractr")
 
+        # channel naming (catch if none or too few)
         image_names <- names(y)
+        nchannel <- terra::nlyr(y)
+        if (is.null(image_names) ||
+            nchannel > 1L && length(unique(image_names)) == 1L) {
+            names(y) <- sprintf("channel_%d", seq_len(nchannel))
+        }
 
         # NSE vars
         coverage_fraction <- NULL